This is a first cut at getting TRACE in CSL working better (with a little bit towards restoration of traceset). The new scheme puts responsibility for tracing on the place in the code where a function is called,where the previous version made it the job of the function that was being called. A consequence of this is that (mostly...) all functions including ones built into the CSL kernel or compiled into C++ as part of Reduce optimisation can be traced provided they are CALLED from bytecoded or interpreted stuff. Calls out from stuff that has been compiled into C++ will not be noticed by the tracer. Because things work this way it should be that all calla made from bytecoded stuff can show what function called the traced function (I find this very helpful indeed. I first trace a low0ish level function I am worried about and I can climb back by observing what called it and tracing that too...). When debugging you may often wish to use bootstrapreduce rather than reduce/redcsl so that this is not messed up by optimisation into C++ code. There should be better trace output associated with some tail-calls. Traceset should report on assignment to fluid or global variable in bytecoded functions (but not to local variables). AT present the interpreter does not do traceset and does not keep track of the currently executing function, but I hope to improve that soon, and feel that where I have go to now is stable enough to be worth exposing to people who may try it out. As always there may be glitches which I hope you will report!
Now a few caveats. The bytestream compiler turns some things (eg CAR CDR, but
a number of other things) into dedicated byte operations. So the tracer can not
see those. Printing trace output (at present) uses some Lisp code, so if you
trace a low level function used in that you are liable to find yourself in
a recursive mess. Eg consider trying to go "(trace '(print))" so that while
trying to tell you that print is being called its gets called again. And as I
have already said several times, code that has been optimised into C is
more tracable than in my previous version, but calkls out from it will not
be notices by the tracer. The compile rcan also expand some recursive functions
into an iterative form, so when you trace them you will not see the nast of
calls that you may have naively expected.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3896 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/newhash.cpp
@@ -399,10 +399,10 @@ void dumptable(LispObject tt, const char *s, bool checkdups)
                 if (ht(h2) == SPID_HASHEMPTY) s4 = " @@@";
             }
             if (h1 != i && h2 != i && h3 != i) s4 = "@@@";
-            printf("%3" PRIuMAX ": [%" PRIu64 "] %s%" PRIuMAX " %s%" PRIuMAX
-                   " %s%" PRIuMAX "%s (%" PRIu64 ")\n",
-                (uintmax_t)i, (uint64_t)k,
-                s1, (uintmax_t)h1, s2, (uintmax_t)h2, s3, (uintmax_t)h3, s4,
+            printf("%3" PRIu64 ": [%" PRIu64 "] %s%" PRIu64 " %s%" PRIu64
+                   " %s%" PRIu64 "%s (%" PRIu64 ")\n",
+                (uint64_t)i, (uint64_t)k,
+                s1, (uint64_t)h1, s2, (uint64_t)h2, s3, (uint64_t)h3, s4,
                 (uintmax_t)vv);
         }
     }
@@ -618,8 +618,8 @@ bool hash_remove(LispObject key)
 
 size_t hash_insert_if_possible(LispObject key)
 {
-    int Qn;
-    int Q[QSIZE];
+    size_t Qn;
+    size_t Q[QSIZE];
     uint64_t h = HASH(key);
     LispObject v1, v2, v3;
     size_t n, n1, n2, n3;
@@ -741,7 +741,7 @@ size_t hash_insert_if_possible(LispObject key)
            (uint64_t)key);
     printf("before unwind Qn=%d\n", Qn);
     dumptable(tt, "Before", false);
-    {   int j;
+    {   size_t j;
         for (j=0; j<Qn; j++) printf("%d: %d\n", j, Q[j]);
     }
 #endif
@@ -749,7 +749,7 @@ size_t hash_insert_if_possible(LispObject key)
 // as to leave a place available where the key I wish to insert can go.
     Qn = Qn/2;
     while (Qn > 3)
-    {  int j = Qn/2;   // parent
+    {  size_t j = Qn/2;   // parent
 #ifdef TRACE
        printf("move %" PRIx64 " from %d to %d\n", ht(Q[j]), Q[j], Q[Qout]);
 #endif
@@ -892,7 +892,15 @@ static void newhash_rehash(LispObject tab, bool after_gc)
 /////////////////////////////////////////////////////////////////////
 
 // This is simplified printing and sends its output to stderr. It is ONLY
-// intended for use while debugging.
+// intended for use while debugging. I will use if when printing trace
+// and backtrace output.. that gives up on having flexible control over
+// diagnostic output or redirecting it to somewhere other than stderr
+// (in non-windowed mode) or the screen (in windowed mode).
+// I want it to be able to cope with looped up structures so I will use
+// and EQ-keyed hash table to detect places where the structure is
+// reentrant.
+// I will also implement a scheme that lets me limit the output from each
+// printed expression to a certain number of lines of output...
 
 static int simple_column = 0;
